Unify text to file and dir functions
